Fixing test script
