+++ graphql/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node index.js",
     "start-dev": "nodemon index.js",
-    "test": "mocha tests"
+    "test": "nyc mocha tests.js"
   },
   "keywords": [],
   "author": "",
